Add crypto scenarios with fractional-quantity backtesting
Quantity and bar volume are now float64 end-to-end so agents can hold
fractional positions (e.g. 0.25 BTC) — the prerequisite for crypto, where
whole units are too coarse (1 BTC can exceed a whole account). The engine
snaps dust positions to flat (qtyEpsilon) so fractional closes net cleanly.

- Fractional quantity: models, engine (weighted-avg + epsilon close),
  trades API (exclusiveMinimum:0), and the bottrade-mcp service.
- Float bar volume everywhere; DB columns stay INTEGER (SQLite affinity
  stores REAL losslessly, no migration).
- Crypto data: GetHistoricalCryptoCandles (Alpaca 24/7 crypto feed),
  CryptoUniverse + IsCryptoSymbol, crypto slippage tiers, crypto routing in
  the hourly ingest job and backfill CLI.
- Four crypto scenarios (BTC/USD benchmark): 2021 blow-off top, FTX collapse,
  spot-ETF sell-the-news, $100k breakout.
- Reference bots, agent-skills.md, SKILL.md, llms.txt, methodology, and
  ARCHITECTURE updated for fractional/crypto.

+++ bottrade-mcp/bottrade.go
@@ -69,7 +69,7 @@ type Run struct {
 type Position struct {
 	RunID     string    `json:"run_id"`
 	Symbol    string    `json:"symbol"`
-	Quantity  int       `json:"quantity"`
+	Quantity  float64   `json:"quantity"`
 	AvgCost   float64   `json:"avg_cost"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
@@ -79,7 +79,7 @@ type QueuedOrder struct {
 	RunID           string    `json:"run_id"`
 	Symbol          string    `json:"symbol"`
 	Side            string    `json:"side"`
-	Quantity        int       `json:"quantity"`
+	Quantity        float64   `json:"quantity"`
 	Reasoning       string    `json:"reasoning,omitempty"`
 	QueuedAt        time.Time `json:"queued_at"`
 	QueuedAtSimTime time.Time `json:"queued_at_sim_time"`
@@ -90,7 +90,7 @@ type Trade struct {
 	RunID         string    `json:"run_id"`
 	Symbol        string    `json:"symbol"`
 	Side          string    `json:"side"`
-	Quantity      int       `json:"quantity"`
+	Quantity      float64   `json:"quantity"`
 	FillPrice     float64   `json:"fill_price"`
 	SlippageBps   int       `json:"slippage_bps"`
 	SimTimeFilled time.Time `json:"sim_time_filled"`
@@ -131,7 +131,7 @@ type MarketBar struct {
 	High   float64 `json:"high"`
 	Low    float64 `json:"low"`
 	Close  float64 `json:"close"`
-	Volume int64   `json:"volume"`
+	Volume float64 `json:"volume"` // fractional for crypto coin volume
 }
 
 type MarketResponse struct {
@@ -171,10 +171,10 @@ type RunResults struct {
 }
 
 type TradeOrder struct {
-	Symbol    string `json:"symbol"`
-	Side      string `json:"side"`
-	Quantity  int    `json:"quantity"`
-	Reasoning string `json:"reasoning,omitempty"`
+	Symbol    string  `json:"symbol"`
+	Side      string  `json:"side"`
+	Quantity  float64 `json:"quantity"`
+	Reasoning string  `json:"reasoning,omitempty"`
 }
 
 type QueuedTradeResult struct {
